# swagger.yml
swagger: "2.0"
info:
  description: "Backend REST API using Node, Express, Sequelize and MySql with Spotify API"
  version: "1.0.1"
  title: "🎹 UMG Spotify Track REST API 🎹"
  
servers:
  - url: https://api.umg.codes/v1/api
    description: Optional server description, e.g. Main (production) server
  - url: https://api.umg.codes/v1/api
    description: Optional server description, e.g. Internal staging server for testing

schemes: 
  - http
  - https

host: 
  - api.umg.codes
 # - umg-mvp.herokuapp.com
 # - localhost:7001

basePath: "/v1/api/isrc"

paths:
  /artists:
    get:
      summary: "Search for artists by Spotify ID or query string"
      description: "Get one artist by Spotify ID or  Get all artist by name matching a query string \n\n `?search=` accepts a text string that will fuzzy match against artist names (Ex 'Daft' => 'Daft Punk', 'The' => 'The Beatles' and 'The White Stripes') \n `?id=` accepts the Spotify assigned ID of an artist (Ex '2cnMpRsOVqtPMfq7YiFE6K' => 'Van Halen')"
      produces:
        - "application/json"
      parameters:
        - in: query
          name: search
          schema: 
            type: string
            example: "Daft"
            summary: ""
        - in: query
          name: id
          schema: 
            type: string
            example: "3yDIp0kaq9EFKe07X1X2rz"
            summary: ""

      responses:
        200:
          description: 200 Success
          schema:
            type: array
            items:
        500:
          description: 500 Internal Server Error
          schema:
            type: object
            items:
    
  /track/{isrcId}:
    get:
      summary: "Get track record"
      description: "Get a track record using ISRC code"
      produces:
        - "application/json"
      parameters:
        - in: path
          name: isrcId
          schema:
            type: string
            example: "GBAYE0601498"
      responses:
        200:
          description: 200 Success
          schema:
            type: object
            items:
        500:
          description: 500 Internal Server Error
          schema:
            type: object
            items:
    post:
      summary: "Create track record"
      description: "Update a track record using ISRC code"
      produces:
        - "application/json"
      parameters:
        - in: path
          name: isrcId
          schema:
            type: string
            example: "GBDUW0100003"
      responses:
        200:
          description: 200 Success
          schema:
            type: object
            items:
        500:
          description: 500 Internal Server Error
          schema:
            type: object
            items:
    